requirements.txt.txt
random
pillow
easygui
rembg
tqdm
pyttsx3
